action buttons, media template tab, terms template tab

+++ webapp/src/org/mig/view/mediators/managers/templates/tabs/TemplateMediaTabMediator.as
@@ -16,6 +16,8 @@ package org.mig.view.mediators.managers.templates.tabs
 	import org.mig.view.events.ContentViewEvent;
 	import org.robotlegs.mvcs.Mediator;
 	
+	import spark.events.TextOperationEvent;
+	
 	public class TemplateMediaTabMediator extends Mediator
 	{
 		[Inject]
@@ -36,6 +38,7 @@ package org.mig.view.mediators.managers.templates.tabs
 		}
 		private function handleViewCreated(event:FlexEvent):void {
 			view.addButton.addEventListener(MouseEvent.CLICK,handleAddButton);
+			view.nameInput.addEventListener(TextOperationEvent.CHANGE,handleNameChange);
 			view.tab.parameters.addEventListener(CollectionEvent.COLLECTION_CHANGE,handleParametersChange);
 		}
 		private function handleAddButton(event:MouseEvent):void {
@@ -49,7 +52,10 @@ package org.mig.view.mediators.managers.templates.tabs
 			param.param3 = view.tab.parameters.length.toString();
 			param.tabid = view.tab.id;
 			param.templateid = view.template.id;
-			
+		}
+		private function handleNameChange(event:TextOperationEvent):void {
+			view.tab.name = view.tab.labelParameter.value = view.tab.labelParameter.updateData.value = view.nameInput.text;
+			eventDispatcher.dispatchEvent(new ContentViewEvent(ContentViewEvent.TEMPLATE_MODIFIED,view.template));
 		}
 		private function handleParametersChange(event:CollectionEvent):void {
 			eventDispatcher.dispatchEvent(new ContentViewEvent(ContentViewEvent.TEMPLATE_MODIFIED,view.template));
@@ -57,28 +63,50 @@ package org.mig.view.mediators.managers.templates.tabs
 		private function handleSubmit(event:ContentViewEvent):void {
 			cudTotal = cudCount = 0;
 			for each(var template:Template in contentModel.templates) {
-				for each(var tab:ContentTab in template.contentTabs) {
-					var param:ContentTabParameter;
-					for each(param in tab.parameters.modifiedItems.source) {
-						cudTotal++;
-						contentService.updateContent(param,contentModel.contentTabsParametersConfig,[]);
-						contentService.addHandlers(handleParamUpdated);
-						contentService.currentToken.tab = tab;
-					}
-					for each(param in tab.parameters.deletedItems.source) {
-						cudTotal++;
-						contentService.deleteContent(param,contentModel.contentTabsParametersConfig);
-						contentService.addHandlers(handleParamDeleted);
-						contentService.currentToken.tab = tab;
-					}
-					for each(param in tab.parameters.newItems.source) {
-						cudTotal++;
-						contentService.createContent(param,contentModel.contentTabsParametersConfig,[],true);	
-						contentService.addHandlers(handleParamCreated);	
-						contentService.currentToken.tab = tab;
+				var tab:ContentTab
+				
+				for each(tab in template.contentTabs) {
+					if(template.contentTabs.newItems.getItemIndex(tab) == -1 && 
+					   template.contentTabs.deletedItems.getItemIndex(tab) == -1) {
+						if(template.contentTabs.modifiedItems.getItemIndex(tab) != -1) {
+							cudTotal++;
+							contentService.updateContent(tab.labelParameter,contentModel.contentTabsParametersConfig,[]);
+							contentService.addHandlers(handleNameChanged);
+							contentService.currentToken.tab = tab;
+							contentService.currentToken.template = template;
+						}
+						var param:ContentTabParameter;
+						for each(param in tab.parameters.modifiedItems.source) {
+							cudTotal++;
+							contentService.updateContent(param,contentModel.contentTabsParametersConfig,[]);
+							contentService.addHandlers(handleParamUpdated);
+							contentService.currentToken.tab = tab;
+						}
+						for each(param in tab.parameters.deletedItems.source) {
+							cudTotal++;
+							contentService.deleteContent(param,contentModel.contentTabsParametersConfig);
+							contentService.addHandlers(handleParamDeleted);
+							contentService.currentToken.tab = tab;
+						}
+						for each(param in tab.parameters.newItems.source) {
+							cudTotal++;
+							contentService.createContent(param,contentModel.contentTabsParametersConfig,[],true);	
+							contentService.addHandlers(handleParamCreated);	
+							contentService.currentToken.tab = tab;
+						}
 					}
 				}
 			}
+		}
+		private function handleNameChanged(data:Object):void {
+			var status:StatusResult = data.result as StatusResult;
+			if(status.success) {
+				var tab:ContentTab = data.token.tab as ContentTab;
+				var template:Template = data.token.template as Template;
+				tab.name = tab.labelParameter.value;
+				template.contentTabs.setItemNotModified(tab);
+				checkCudCount();
+			}				
 		}
 		private function handleParamDeleted(data:Object):void {
 			var status:StatusResult = data.result as StatusResult;
